scripts/requirements.txt
paho-mqtt==1.6.1
kafka-python
requests